Add in Graphics and Code Preview
Now at the bottom of the tutorial page there will be a code preview of the sort function
that you have created. We have seen a couple other animation apps that have this
functionality so we wanted to add it in. The other piece is that when the sort function
changes, the code is printed to the console. This makes sure it's really easy to copy
if the developer wants to.