Add fastlane to build process
Adds in fastlane support so that integration with a public CI server will be easier. The main
lane setup is the `run_tests` lane which will run all of the unit tests that are currently
built into Spruce. As we add more and more tests, this functionality should remain the same.

If we wanted to actually build and sign the example project ipa, then this is where we would
also do that. We may want to consider adding in a step so that it does build the example
project, just to make sure that it builds. Will discuss with Biscuit to see more of this.