Add time labels to scrub bars
In the example app we have scrub bars for the duration and delay properties of the animation. Currently,
there are no labels for what their values are. To make it easier to see exactly what the user is setting,
we added labels that describe the min/max values of the scrub bar.