Fix delay slider value issue
Turns out the Delay slider was disconnected for the right IBAction outlet. Needed to wire that back up correctly
and then also change the number of decimals that were being displayed in the code. Since the delay is such a tiny
double, the value may easily be less than 0.01, so 3 decimals were needed.