Respond to PR Comments
Since there were quite a few comments, it made it easier to put them all in one pull request. This way we could both avoid
merge conflicts and try to get everything so that nothing is left out. In this PR, I renamed all of the `Spruce` methods
so that they don't include the prefix.