Update naming scheme to prefix extension methods
Since `Swift` extensions are essentially categories in objective-c, we need to make sure that
we prefix each of the methods with an identifying symbol. To do so, we changed methods such
as `spruceUp` to `spruce_up` so that the prefix is `spruce_`. Anywhere we have a extended method
this prefix is attached.

The question is whether or not `spruce_` is a good prefix. Whether we should keep it or change it?
And with this new prefix should we change any of the method names so that they make more sense.

We were playing around with the idea of `spruce_animate` instead of `spruce_up`. Let us know your
thoughts :)