Update code to support new timing functions
With the new `TimingFunction` protocol being added, we needed to make sure that it works properly with the example app. This will be a breaking change for some functions but we are going to submit another PR on top of this that should minimize the breaking nature of this change.