Update Sort Function unit tests
Since we are changing the concept of a `SortFunction` the unit tests have to change with it. This means that rather than comparing the timing we would compare the weights returned by the `SortFunction`.

However, this changes the concept of the unit tests entirely so we also need to add in unit tests for the `TimingFunction`'s as well. To make sure that the refactor returns the same values, we would also need a combination of unit tests. This means the sort function would have to return the same values if we passed the weights into a standard timing function.